requirements.txt
hidapi
numpy